PMC: Fix build errors on Linux

+++ src/tools/pegasus-mpi-cluster/tools.cpp
@@ -239,7 +239,7 @@ bool is_executable(const string &file) {
     return false;
 }
 
-int read_file(const string &file, char *buf, size_t size) {
+size_t read_file(const string &file, char *buf, size_t size) {
     // Invalid path
     if (file.size() == 0) {
         errno = ENOENT;
